backend/requirements.txt
# FastAPI
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9

# Validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# Security
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# Utils
python-dotenv==1.0.0
